tests/requirements.txt
# Integration Test Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
httpx==0.25.2
boto3==1.34.10
python-jose[cryptography]==3.3.0
pyyaml==6.0.1
